apps/adk-orchestrator/requirements.txt
fastapi>=0.110
uvicorn>=0.27
pydantic>=2.0
pydantic-settings>=2.2
PyYAML>=6.0
aiosqlite>=0.20